+++ src/yetty/terminal.h
@@ -29,8 +29,8 @@ public:
     virtual void setViewport(float x, float y, float width, float height) = 0;
 
     // Cell size (from font metrics * zoom)
-    virtual uint32_t getCellWidth() const = 0;
-    virtual uint32_t getCellHeight() const = 0;
+    virtual float getCellWidth() const = 0;
+    virtual float getCellHeight() const = 0;
 
 protected:
     Terminal() = default;
